cleanup enum support for typescript
