--- src/node-swagger-generator/templates/typescript/angular/2.0/_definition.hbs
@@ -1,18 +1,0 @@
-
-{{#if description}}
-/**
-*{{description}}
-*/
-{{/if}}
-export interface {{name}} {{#ancestor}}extends {{name}}{{/ancestor}} {
-
-{{#properties}}
-  {{#if description}}
-    /**
-    *{{description}}
-    */
-  {{/if}}
-  {{name}} : {{> type }};
-
-{{/properties}}
-}
