--- src/node-swagger-generator/templates/typescript/angular/2.0/_type.hbs
@@ -1,1 +1,0 @@
-{{> typeCore type}}
