--- src/node-swagger-generator/templates/typescript/angular/2.0/_typeCore.hbs
@@ -1,6 +1,0 @@
-{{#isBuiltin}}{{name}}{{/isBuiltin}}{{#isDefinition}}{{name}}{{/isDefinition}}{{#isAnonymous}} {
-    {{#definition.properties}}
-        /*{{description}}*/
-        {{name}} : {{> typeCore type }};
-    {{/definition.properties}}
-    }{{/isAnonymous}}{{#isArray}}{{~> typeCore itemType ~}}[]{{/isArray}}
