--- src/node-swagger-generator/templates/typescript/angular/2.0/client.hbs
@@ -1,10 +1,0 @@
-import {Http} from "@angular/http";
-
-export class {{clientName}}{{#if options.generateInterface}} implements I{{clientName}}{{/if}} {
-
-    constructor(
-        private http: Http
-    ) {
-        
-    }
-}
