--- src/node-swagger-generator/templates/typescript/angular/2.0/definition.hbs
@@ -1,1 +1,0 @@
-{{> definition data}}
