--- src/node-swagger-generator/templates/typescript/angular/2.0/template.json
@@ -1,27 +1,0 @@
-{
-  "name": "typescript-angular-2.",
-  "language": {
-    "name": "typescript",
-    "defaultExtension": ".ts"
-  },
-  "dependencies": {},
-  "definitionFilters": [
-    "arrayName",
-    "pascalCaseName"
-  ],
-  "operationFilters": [
-    "camlCaseName",
-    "camlCaseArgumentName"
-  ],
-  "modes": {
-    "singleFile": {
-      "entries": [
-        {
-          "template": "singleFile.hbs",
-          "naming": "{{options.clientName}}.ts",
-          "selector": ""
-        }
-      ]
-    }
-  }
-}
