+++ templates/typescript/angular/1.5/_typeCore.hbs
@@ -1,6 +1,6 @@
-{{#isBuiltin}}{{name}}{{/isBuiltin}}{{#isDefinition}}{{name}}{{/isDefinition}}{{#isAnonymous}} {
+{{#isBuiltin}}{{name}}{{/isBuiltin}}{{#isDefinition}}{{#if import}}{{import}}.{{/if}}{{name}}{{/isDefinition}}{{#isAnonymous}} {
     {{#definition.properties}}
         /*{{description}}*/
-        {{name}} : {{> type }};
+        {{name}} : {{> typeCore type import=import }};
     {{/definition.properties}}
-}{{/isAnonymous}}{{#isArray}}{{> typeCore itemType}}[]{{/isArray}}
+    }{{/isAnonymous}}{{#isArray}}{{~> typeCore itemType import=import ~}}[]{{/isArray}}{{#isEnum}}{{#enumValues}}{{intOrString .}}{{#unless @last}} | {{/unless}}{{/enumValues}}{{/isEnum}}
