+++ templates/typescript/angular/4.0/_typeCore.hbs
@@ -3,4 +3,4 @@
         /*{{description}}*/
         {{name}} : {{> typeCore type import=import }};
     {{/definition.properties}}
-    }{{/isAnonymous}}{{#isArray}}{{~> typeCore itemType import=import ~}}[]{{/isArray}}
+    }{{/isAnonymous}}{{#isArray}}{{~> typeCore itemType import=import ~}}[]{{/isArray}}{{#isEnum}}{{#enumValues}}{{intOrString .}}{{#unless @last}} | {{/unless}}{{/enumValues}}{{/isEnum}}
